improved and added tests

+++ STHTTPRequest2/Demo Project/Unit Tests/STHTTPRequestTests.m
@@ -79,7 +79,7 @@
     
     STHTTPRequest *r = [STHTTPRequest requestWithURLString:@"http://www.google.com"];
     
-    r.downloadProgressBlock = ^(NSData *data, NSUInteger totalBytesReceived, long long totalBytesExpectedToReceive) {
+    r.downloadProgressBlock = ^(NSData *data, int64_t totalBytesReceived, int64_t totalBytesExpectedToReceive) {
         NSLog(@"-- %@", data);
     };
